requirements.txt
dj-rest-auth==6.0.0
Django==5.1.1
django-allauth==0.61.1
djangorestframework==3.15.2
pillow==10.4.0
requests==2.32.3